requirements.txt
google-ads>=29.0.0
google-api-python-client>=2.100.0
google-auth>=2.20.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
pandas>=2.0.0
python-dotenv>=1.0.0
